requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
django-cors-headers==4.3.1
gunicorn==21.2.0
Pillow==10.1.0
django-filter==23.4
drf-yasg==1.21.7
channels==4.0.0
